requirements.txt
absl_py==1.0.0
keras==2.10.0
matplotlib==3.5.1
numpy==1.22.1
opencv_python==4.5.5.62
pandas==1.4.0
pycm==3.4
scikit_learn==1.0.2
scipy==1.7.3
tensorflow==2.10.0
tensorflow_datasets==4.5.2
git+https://github.com/google/qkeras.git==0.9.0